requirements.txt
Flask>=2.0.0
python-docx>=0.8.0
docx2txt>=0.8
faiss-cpu>=1.7.0
openai>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
plotly>=5.0.0
sentence-transformers>=2.0.0
numpy>=1.20.0
Pillow>=8.0.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0
beautifulsoup4>=4.10.0
lxml>=4.6.0
openpyxl>=3.0.0